feat: updated `run.sh` with new flags.

+++ run.sh
@@ -48,17 +48,21 @@ fi
 set -x
 
 # Calculate ego vehicle masks
-echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
-echo "~~~~~~~~~~~~~ Calculating ego vehicle masks ~~~~~~~~~~~~~"
-echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
-rm -rf $EGO_MASK_PATH
-mkdir $EGO_MASK_PATH
-python infer.py $EGO_MODEL_CONFIG $1 $EGO_MASK_PATH --only_ego_vehicle --n_skip_frames -2 --out_format png --no_tqdm $EGO_INPUT_SHAPE_ARG $EGO_BATCH_SIZE_ARG #  --n_skip_frames -2 - take one frame per 2 seconds
+APPLY_EGO_MASK_FROM=""
+if [ $SKIP_EGO_VEHICLE == 0 ]; then
+  echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
+  echo "~~~~~~~~~~~~~ Calculating ego vehicle masks ~~~~~~~~~~~~~"
+  echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
+  rm -rf $EGO_MASK_PATH
+  mkdir $EGO_MASK_PATH
+  python infer.py $EGO_MODEL_CONFIG $1 $EGO_MASK_PATH --only_ego_vehicle --n_skip_frames -2 --out_format png --no_tqdm $EGO_INPUT_SHAPE_ARG $EGO_BATCH_SIZE_ARG #  --n_skip_frames -2 - take one frame per 2 seconds
+  APPLY_EGO_MASK_FROM="--apply_ego_mask_from $EGO_MASK_PATH"
+fi
 # Calculate segmentation masks
 echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
 echo "~~~~~~~~~~~~~ Calculating segmentation masks ~~~~~~~~~~~~~"
 echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
-python infer.py $MAIN_MODEL_CONFIG $1 $2 --apply_ego_mask_from $EGO_MASK_PATH --no_tqdm $MAIN_INPUT_SHAPE_ARG $MAIN_BATCH_SIZE_ARG
+python infer.py $MAIN_MODEL_CONFIG $1 $2 $APPLY_EGO_MASK_FROM --no_tqdm $MAIN_INPUT_SHAPE_ARG $MAIN_BATCH_SIZE_ARG
 
 # Other onnx models
 # python infer.py configs/onnx-deeplabv3plus-r18d-dynamic.yaml $1 $2 --apply_ego_mask_from $EGO_MASK_PATH
